requirements.txt
openai
faiss-cpu
numpy
requests
beautifulsoup4
python-dotenv
fastapi
uvicorn
pydantic 